+++ lime-loader/src/main/java/com/here/genium/loader/AntlrLimedocBuilder.kt
@@ -20,78 +20,61 @@
 package com.here.genium.loader
 
 import com.here.genium.antlr.LimedocParser
-import com.here.genium.antlr.LimedocParserListener
-import org.antlr.v4.runtime.ParserRuleContext
-import org.antlr.v4.runtime.tree.ErrorNode
-import org.antlr.v4.runtime.tree.TerminalNode
+import com.here.genium.antlr.LimedocParserBaseListener
+import com.here.genium.model.lime.LimeComment
 
-internal class AntlrLimedocBuilder : LimedocParserListener {
+internal class AntlrLimedocBuilder : LimedocParserBaseListener() {
 
-    private val commentsCollector = mutableMapOf<Pair<String, String>, String>()
+    private val commentsCollector = mutableMapOf<Pair<String, String>, LimeComment>()
+    private val contentCollector = mutableListOf<Pair<String, String>>()
 
     val result
-        get() = LimeStructuredComment(commentsCollector[Pair("", "")] ?: "", commentsCollector)
+        get() = LimeStructuredComment(
+            commentsCollector[Pair("", "")] ?: LimeComment(),
+            commentsCollector
+        )
 
     // Overrides
 
     override fun exitDescription(ctx: LimedocParser.DescriptionContext) {
-        commentsCollector[Pair("", "")] = ctx.text.trim()
+        commentsCollector[Pair("", "")] = LimeComment(contentCollector.toList())
+        contentCollector.clear()
+    }
+
+    override fun enterDescriptionMoreLines(ctx: LimedocParser.DescriptionMoreLinesContext) {
+        contentCollector += "" to "\n".repeat(ctx.NL().size)
+    }
+
+    override fun exitDecriptionFirstWord(ctx: LimedocParser.DecriptionFirstWordContext) {
+        contentCollector += when {
+            ctx.inlineTag() != null -> convertInlineTag(ctx.inlineTag())
+            else -> "" to ctx.text
+        }
+    }
+
+    override fun exitDescriptionContent(ctx: LimedocParser.DescriptionContentContext) {
+        contentCollector += when {
+            ctx.inlineTag() != null -> convertInlineTag(ctx.inlineTag())
+            else -> "" to ctx.text
+        }
     }
 
     override fun exitBlockTag(ctx: LimedocParser.BlockTagContext) {
-        val tagName = ctx.blockTagName().text
+        val tagName = ctx.tagName().text
         val tagParameter = ctx.blockTagParameter()?.text ?: ""
-        commentsCollector[Pair(tagName, tagParameter)] =
-            ctx.blockTagContent().joinToString("") { it.text }.trim()
+        commentsCollector[Pair(tagName, tagParameter)] = LimeComment(contentCollector.toList())
+        contentCollector.clear()
     }
 
-    // Empty overrides
+    override fun exitBlockTagContent(ctx: LimedocParser.BlockTagContentContext) {
+        contentCollector += when {
+            ctx.inlineTag() != null -> convertInlineTag(ctx.inlineTag())
+            else -> "" to ctx.text
+        }
+    }
 
-    override fun enterDocumentation(ctx: LimedocParser.DocumentationContext) {}
+    // Private functions
 
-    override fun exitDocumentation(ctx: LimedocParser.DocumentationContext) {}
-
-    override fun enterSkipWS(ctx: LimedocParser.SkipWSContext) {}
-
-    override fun enterDocumentationContent(ctx: LimedocParser.DocumentationContentContext) {}
-
-    override fun exitDocumentationContent(ctx: LimedocParser.DocumentationContentContext) {}
-
-    override fun enterDescription(ctx: LimedocParser.DescriptionContext) {}
-
-    override fun enterDescriptionLine(ctx: LimedocParser.DescriptionLineContext) {}
-
-    override fun visitErrorNode(node: ErrorNode?) {}
-
-    override fun exitDescriptionLine(ctx: LimedocParser.DescriptionLineContext) {}
-
-    override fun enterDescriptionText(ctx: LimedocParser.DescriptionTextContext?) {}
-
-    override fun exitDescriptionText(ctx: LimedocParser.DescriptionTextContext?) {}
-
-    override fun visitTerminal(node: TerminalNode?) {}
-
-    override fun exitSkipWS(ctx: LimedocParser.SkipWSContext) {}
-
-    override fun enterTagSection(ctx: LimedocParser.TagSectionContext) {}
-
-    override fun exitTagSection(ctx: LimedocParser.TagSectionContext?) {}
-
-    override fun enterBlockTag(ctx: LimedocParser.BlockTagContext) {}
-
-    override fun enterBlockTagName(ctx: LimedocParser.BlockTagNameContext) {}
-
-    override fun exitBlockTagName(ctx: LimedocParser.BlockTagNameContext) {}
-
-    override fun enterBlockTagParameter(ctx: LimedocParser.BlockTagParameterContext?) {}
-
-    override fun exitBlockTagParameter(ctx: LimedocParser.BlockTagParameterContext?) {}
-
-    override fun enterBlockTagContent(ctx: LimedocParser.BlockTagContentContext) {}
-
-    override fun exitBlockTagContent(ctx: LimedocParser.BlockTagContentContext) {}
-
-    override fun enterEveryRule(ctx: ParserRuleContext) {}
-
-    override fun exitEveryRule(ctx: ParserRuleContext) {}
+    private fun convertInlineTag(inlineTag: LimedocParser.InlineTagContext) =
+        inlineTag.tagName().text to inlineTag.inlineTagContent().joinToString("") { it.text }
 }